Add new float handling to solve last FP issues, allow decimal power and logistics

+++ Reika/SatisfactoryPlanner/Data/Objects/ResourceSupplies/DroneStation.java
@@ -21,7 +21,7 @@ public class DroneStation extends LogisticSupply<DroneStation, Item> {
 
 	private DroneStation(JSONObject obj) {
 		this((Item)Database.lookupItem(obj.getString("item")));
-		this.setAmount(obj.getInt("amount"));
+		this.setAmount(obj.getDouble("amount"));
 	}
 
 	@Override
